Modified error

+++ FundusImageDownload.py
@@ -1,14 +1,6 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# # Retinal Fundus Image
-# - 총 5개의 Data Set (4.85G)
-# - 파이썬 파일과 동일한 폴더에 다운로드 및 정리
-#   - 데이터 폴더 : 'Train'
-#   - 데이터 관련 CSV파일 : 'Train/Train.tsv'
-#   - 분류 결과 설명 파일 : 'Train/Traininfo.html'
-# - 데이터 이미지 확장자는 jpg, bmp
-
 import http
 import zipfile
 import urllib.request
@@ -25,19 +17,21 @@ import time
 
 # 데이터셋 출처&다운로드 링크
 link_File = 'UrlList.csv'
-arg = open(link_File, 'w', encoding='utf=8', newline="")
-wr = csv.writer(arg)
-wr.writerow(['STARE', 'http://cecas.clemson.edu/~ahoover/stare/images/all-images.zip', 'STARE_Project.zip'])
-wr.writerow(['STARE', "http://cecas.clemson.edu/~ahoover/stare/diagnoses/all-mg-codes.txt", 'STARE/info.txt'])
-wr.writerow(['STARE', "http://cecas.clemson.edu/~ahoover/stare/diagnoses/diagnoses.html", 'STARE/diagnoses.html'])
-wr.writerow(['csfau', 'https://www5.cs.fau.de/fileadmin/research/datasets/fundus-images/all.zip', 'csfau.zip'])
-wr.writerow(['csfau', "https://www5.cs.fau.de/fileadmin/research/datasets/fundus-images/allQuality.zip",
-            'csfau/test.zip'])
-wr.writerow(['Yiweichen04', 'https://github.com/yiweichen04/retina_dataset/archive/master.zip', 'yiweichen04.zip'])
-wr.writerow(['ACRIMA', 'https://ndownloader.figshare.com/files/14137700', 'ACRIMA.zip'])
-wr.writerow(['Messidor-2', "https://www.ceos-systems.com/file-sharing/IMAGES.zip.001", 'Messidor-2.zip.001'])
-wr.writerow(['Messidor-2', 'https://codeload.github.com/Yeo-Jun-Choi/download/zip/main', 'Messidor-2/messidor_data.zip'])
-arg.close()
+if not os.path.isfile(link_File):
+    arg = open(link_File, 'w', encoding='utf=8', newline="")
+    wr = csv.writer(arg)
+    wr.writerow(['STARE', 'http://cecas.clemson.edu/~ahoover/stare/images/all-images.zip', 'STARE_Project.zip'])
+    wr.writerow(['STARE', "http://cecas.clemson.edu/~ahoover/stare/diagnoses/all-mg-codes.txt", 'STARE/info.txt'])
+    wr.writerow(['STARE', "http://cecas.clemson.edu/~ahoover/stare/diagnoses/diagnoses.html", 'STARE/diagnoses.html'])
+    wr.writerow(['csfau', 'https://www5.cs.fau.de/fileadmin/research/datasets/fundus-images/all.zip', 'csfau.zip'])
+    wr.writerow(['csfau', "https://www5.cs.fau.de/fileadmin/research/datasets/fundus-images/allQuality.zip",
+                 'csfau/test.zip'])
+    wr.writerow(['Yiweichen04', 'https://github.com/yiweichen04/retina_dataset/archive/master.zip', 'yiweichen04.zip'])
+    wr.writerow(['ACRIMA', 'https://ndownloader.figshare.com/files/14137700', 'ACRIMA.zip'])
+    wr.writerow(['Messidor-2', "https://www.ceos-systems.com/file-sharing/IMAGES.zip.001", 'Messidor-2.zip.001'])
+    wr.writerow(['Messidor-2', 'https://codeload.github.com/Yeo-Jun-Choi/download/zip/main', 'Messidor-2/messidor_data.zip'])
+    arg.close()
+
 
 # sys.argv로 추가적인 데이터 셋 받을 때 처리 정의
 while (True):
@@ -55,7 +49,7 @@ while (True):
     else:
         break
 
-# 데이터 개수 세기 위한 변수
+# 각기 다른 데이터 이름을 하나로 통합할 때, 순서를 위해 전역변수 k를 사용
 k = 1
 
 # 폴더 생성 함수
@@ -65,16 +59,13 @@ def direct(path):
     except FileExistsError:
         pass
 
-# ## Dataset 처리 관련 변수 정의
-
+# Data 분류에 사용할 병명 dict형태로 정리
 from collections import defaultdict
-
 def invert_dictionary(obj):
     inv_obj = defaultdict(list)
     for key, value in obj.items():
         inv_obj[value].append(key)
     return dict(inv_obj)
-
 recode = {
     "Normal": 0,
     "diabetic_retinopathy": 15,
@@ -98,8 +89,7 @@ recode = {
 }
 r_recode = invert_dictionary(recode)
 
-
-# Dataset 처리 관련 폴더 생성
+# 분류데이터 설명 Html파일 작성
 def Make_Info_Html():
     namecol = 'Disease Name'
     df = pd.read_csv('Train/Train.tsv', sep='\t', header=None)
@@ -120,32 +110,17 @@ def Make_Info_Html():
     with open('Train/Traininfo.html', 'w', newline="") as Traininfo:
         Traininfo.write(df1)
         Traininfo.write(html)
+
 try:
     os.makedirs(os.path.join("Train"))
 except FileExistsError:
     pass
-direct("diabetic_retinopathy")
-direct("glaucomatous")
-direct("Normal")
-direct("cataract")
-direct("retina_disease")
-direct("Hollenhorst_Emboli")
-direct("Branch_Retinal_Artery_Occlusion")
-direct("Cilio-Retinal_Artery_Occlusion")
-direct("Branch_Retinal_Vein_Occlusion")
-direct("Central_Retinal_Vein_Occlusion")
-direct("Hemi-Central_Retinal_Vein_Occlusion")
-direct("Background_Diabetic_Retinopathy")
-direct("Proliferative_Diabetic_Retinopathy")
-direct("Arteriosclerotic_Retinopathy")
-direct("Hypertensive_Retinopathy")
-direct("Coats")
-direct("Macroaneurism")
-direct("Choroidal_Neovascularization")
-direct("Diabetic_Macular_Edema")
+
+# Dataset 처리 관련 폴더 생성
+for i in range(len(r_recode)):
+    direct(r_recode[i][0])
 
 # 전체 데이터셋 설명 tsv파일 생성
-
 writedata = open('Train/Train.tsv', 'w', newline="")
 fieldnames = ['Filename', 'number', 'desease']
 wrt = csv.writer(writedata, fieldnames, delimiter="\t")
@@ -166,12 +141,12 @@ class data_download:
             all_ForderList.append(patha)
         if not all_ForderList:
             filelist += [file for file in allFile if
-                         file.upper().endswith(".JPG") or file.upper().endswith(".BMP") or file.upper().endswith(".PNG")]
+                         file.upper().endswith(".JPG") or file.upper().endswith(".BMP") or file.upper().endswith(".PNG") or file.upper().endswith(".PPM")]
         else:
             for forder in all_ForderList:
                 file_Forder = forder+'\*'
                 filelist += [file for file in list(filter(os.path.isfile, glob.glob(file_Forder))) if
-                         file.upper().endswith(".JPG") or file.upper().endswith(".BMP") or file.upper().endswith(".PNG")]
+                         file.upper().endswith(".JPG") or file.upper().endswith(".BMP") or file.upper().endswith(".PNG") or file.upper().endswith(".PPM")]
         return filelist
 
     # 데이터를 'Train'폴더에 통합. (overlap : 중복 허용 || data : data설명 csv, tsv, txt파일)
@@ -199,6 +174,7 @@ class data_download:
                     if(datanum == 14):
                         continue
                     self.copydata(i, j, "Train/" + r_recode[datanum][0], recode)
+
                 except ValueError:
                     pass
         global k
@@ -225,6 +201,7 @@ class data_download:
                 j = j.upper()
                 divide_first_type(data, i, j, num)
                 num += 1
+
         elif (overlap and isinstance(data, list)):
             for i in filelist:
                 x = i.split("\\")
@@ -252,14 +229,18 @@ class data_download:
                 else:
                     self.copydata(i, j, "Train/Normal", recode)
     # 데이터 관련 정보 전처리. csv, tsv파일이 있으면 infodata=True
+
     def Comb(self, path):
         filelist = self.return_allImage(path)
         for i in range(len(filelist)):
             oldname = filelist[i]
             if ('ppm' in oldname):
                 newname = os.path.splitext(oldname)[0] + '.bmp'
+                if(os.path.isfile(path+'/'+newname)):
+                    return
                 image = Image.open(oldname)
                 image.save(newname, 'BMP')
+                filelist[i] = newname
 
         for n_path, dirs, files in os.walk(path):
             for filename in files:
@@ -386,8 +367,6 @@ class data_download:
             obj[num][0].Comb(obj[num][0].name)
             pass
 
-    #데이터 설명 파일(HTML) 생성
-    # Messidor을 해야함
 
 def main():
     f = open(link_File, 'r', encoding='utf-8',newline="")
@@ -405,6 +384,4 @@ def main():
     time.sleep(3)
     writedata.close()
     Make_Info_Html()
-
-
 main()
